OutputMutable: clarify the purpose
I want the intent of this interface to be that the output doubles as
an input -- actually, the first input, even before the unary/binary/etc.
inputs from the perspective of the "vanilla" ops framework.

This means we can reason about whether to treat a special typed output
parameter gleaned from somewhere else (e.g., from a Nil's generic type)
as an input, or a pure output, without hardcoding function/computer/etc.